MUTATION: prove the two database-backed guards can fail
Throwaway. Drops community_kwh from the metering IS DISTINCT FROM guard, and
points the billing_periods community_id cast at a data_type that never matches,
so neither guarded branch fires. Both new tests must go red in ci/test.

+++ store/schema.py
@@ -627,7 +627,7 @@ def create_tables():
                         SELECT 1 FROM information_schema.columns
                         WHERE table_name = 'billing_periods'
                             AND column_name = 'community_id'
-                            AND data_type = 'integer'
+                            AND data_type = 'bigint'
                     ) THEN
                         ALTER TABLE billing_periods
                             ALTER COLUMN community_id TYPE VARCHAR(64);
